Rename divesity_med_qa
