+++ docs/source/probes/diversity_med_qa.rst
@@ -1,0 +1,1 @@
+.. mdinclude:: ../../../src/gender_bench/probes/diversity_med_qa/README.md
