--- docs/source/probes/diversitymedqa.rst
@@ -1,1 +1,0 @@
-.. mdinclude:: ../../../src/gender_bench/probes/diversitymedqa/README.md
