+++ src/gender_bench/probes/__init__.py
@@ -6,7 +6,9 @@ from gender_bench.probes.direct.direct_probe import DirectProbe
 from gender_bench.probes.discrimination_tamkin.discrimination_tamkin_probe import (
     DiscriminationTamkinProbe,
 )
-from gender_bench.probes.diversitymedqa.diversitymedqa_probe import DiversityMedQaProbe
+from gender_bench.probes.diversity_med_qa.diversity_med_qa_probe import (
+    DiversityMedQaProbe,
+)
 from gender_bench.probes.dreaddit.dreaddit_probe import DreadditProbe
 from gender_bench.probes.gest.gest_probe import GestProbe
 from gender_bench.probes.gest_creative.gest_creative_probe import GestCreativeProbe
