+++ src/gender_bench/resources/diversity_med_qa/README.md
@@ -6,6 +6,6 @@ its original MIT license is included in `LICENSE_MedQA`. The DiversityMedQa
 dataset was downloaded from [huggingface.com](https://huggingface.co/datasets/Rajat1212/DiversityMedQA),
 its original CC0 1.0 Universal license is included in `LICENSE_DivesityMedQA`.
 
-The male and female sentences in our dataset (`diversitymedqa_extracted.csv`)
+The male and female sentences in our dataset (`diversity_med_qa_extracted.csv`)
 are extracted from `DiversityMedQA` dataset, while the options are extracted
 from the `MedQA` dataset.
